+++ AFKScape/Assets/Scripts/Drop Table/DropTable.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using System;
+using System.Numerics;
 
 [Serializable]
 public class DropTable
@@ -47,11 +48,11 @@ public class DropTable
         lootItems = new List<Loot>() { new Loot(0) };
     }
 
-    public virtual void RollTable(Dictionary<long, int> dropTableDict)
+    public virtual void RollTable(Dictionary<long, BigInteger> dropTableDict)
     {
     }
 
-    public virtual void RollTable(Dictionary<long, int> dropTableDict, int skillLevel) //is this the way to do it?
+    public virtual void RollTable(Dictionary<long, BigInteger> dropTableDict, int skillLevel) //is this the way to do it?
     {
     }
