+++ AFKScape/Assets/Scripts/Storage/Inventory.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 
 public class Inventory : Storage
@@ -9,12 +10,12 @@ public class Inventory : Storage
         totalSlots = 28;
     }
 
-    public override void RaiseItemAddedEvent(long id, int amount, int amounDiff)
+    public override void RaiseItemAddedEvent(long id, BigInteger amount, BigInteger amounDiff)
     {
         EventManager.Instance.InvItemAdded(id, amount);
     }
 
-    public override void RaiseItemRemovedEvent(long id, int amount, int amounDiff)
+    public override void RaiseItemRemovedEvent(long id, BigInteger amount, BigInteger amounDiff)
     {
         EventManager.Instance.InvItemRemoved(id, amount);
     }
